Backend/requirements.txt
# FastAPI核心
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# 数据库
pymysql==1.1.0
aiomysql==0.2.0
DBUtils==3.1.0
cryptography==42.0.5

# 向量数据库
chromadb==0.4.24
sentence-transformers==2.2.2

# 文档处理
pypdf2==3.0.1
python-docx==1.1.2
beautifulsoup4==4.12.2
lxml==4.9.3
markdown==3.5.1

# 文本处理
# 使用兼容的 langchain 版本组合
langchain==0.1.16
langchain-core==0.1.52
langchain-community==0.0.38
langchain-text-splitters==0.0.1

# 嵌入模型
torch>=2.3.0
transformers>=4.40.0
bitsandbytes>=0.48.0  # 升级到0.48+以支持4-bit量化与device_map="auto"的配合
accelerate>=0.25.0

# 工具库
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pyyaml==6.0.1
aiofiles==23.2.1

# 日志和监控
loguru==0.7.2

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
